fixed initial fetch of the groups, fixed required second click to confirm form action, added copy and mailto features to form info
