+++ src/index.css
@@ -2,8 +2,8 @@
 
 :root {
 	--accent: #654dc4;
+	--accentLowOpacity: #654dc456;
 	--light: #f6f5f8;
-	--blackLowOpacity: #211f265b;
 	--black: #211f26;
 	--white: #ffffff;
